rename: 패키지 분할
- com.music.entity 경로에 선언된 constants 패키지를 com.music으로 변경
- StreamingLog를 Entity 패키지에서 document 패키지로 이동
