+++ core/src/main/java/com/music/repository/MusicChartRepository.java
@@ -1,7 +1,7 @@
 package com.music.repository;
 
 import com.music.eneity.MusicChart;
-import com.music.eneity.constants.ChartType;
+import com.music.constants.ChartType;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
