+++ core/src/main/java/com/music/repository/MusicRepository.java
@@ -1,7 +1,7 @@
 package com.music.repository;
 
 import com.music.eneity.Music;
-import com.music.eneity.constants.ReleaseStatus;
+import com.music.constants.ReleaseStatus;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
