+++ core/src/main/java/com/music/repository/StreamingLogRepository.java
@@ -1,6 +1,6 @@
 package com.music.repository;
 
-import com.music.eneity.StreamingLog;
+import com.music.document.StreamingLog;
 import java.util.Optional;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
