+++ music/src/test/java/com/music/service/MusicStreamingServiceTest.java
@@ -5,8 +5,8 @@ import com.music.streaming.dto.LyricsResponse;
 import com.music.streaming.dto.StreamResponse;
 import com.music.eneity.Lyrics;
 import com.music.eneity.Music;
-import com.music.eneity.constants.LyricsFormat;
-import com.music.eneity.constants.ReleaseStatus;
+import com.music.constants.LyricsFormat;
+import com.music.constants.ReleaseStatus;
 import com.music.repository.LyricsRepository;
 import com.music.repository.MusicRepository;
 import com.music.streaming.service.MusicStreamingService;
