+++ statistics/src/main/java/com/music/chart/convert/ChartTypeConvert.java
@@ -1,6 +1,6 @@
 package com.music.chart.convert;
 
-import com.music.eneity.constants.ChartType;
+import com.music.constants.ChartType;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
