+++ statistics/src/main/java/com/music/chart/service/MusicChartService.java
@@ -5,11 +5,10 @@ import com.music.chart.period.ChartPeriod;
 import com.music.chart.period.ChartPeriodFactory;
 import com.music.chart.util.ChartAggregationBuilder;
 import com.music.chart.util.ChartDataConverter;
-import com.music.chart.util.MusicChartValidator;
 import com.music.eneity.Music;
 import com.music.chart.dto.AggregationResult;
-import com.music.eneity.StreamingLog;
-import com.music.eneity.constants.ChartType;
+import com.music.document.StreamingLog;
+import com.music.constants.ChartType;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
